Clean up helpers refactor
