+++ packages/storage-plus/src/lib.rs
@@ -1,4 +1,5 @@
 mod helpers;
+mod iter_helpers;
 mod keys;
 mod map;
 mod path;
